prune upstream, fix tests and linting
