--- examples/http-jwks-example/main.go
@@ -1,51 +1,0 @@
-package main
-
-import (
-	"encoding/json"
-	"log"
-	"net/http"
-	"net/url"
-	"time"
-
-	"github.com/auth0/go-jwt-middleware/v2"
-	"github.com/auth0/go-jwt-middleware/v2/jwks"
-	"github.com/auth0/go-jwt-middleware/v2/validator"
-)
-
-var handler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-	claims := r.Context().Value(jwtmiddleware.ContextKey{}).(*validator.ValidatedClaims)
-
-	payload, err := json.Marshal(claims)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	w.Write(payload)
-})
-
-func main() {
-	issuerURL, err := url.Parse("https://<your tenant domain>/")
-	if err != nil {
-		log.Fatalf("failed to parse the issuer url: %v", err)
-	}
-
-	provider := jwks.NewCachingProvider(issuerURL, 5*time.Minute)
-
-	// Set up the validator.
-	jwtValidator, err := validator.New(
-		provider.KeyFunc,
-		validator.RS256,
-		issuerURL.String(),
-		[]string{"<your api identifier>"},
-	)
-	if err != nil {
-		log.Fatalf("failed to set up the validator: %v", err)
-	}
-
-	// Set up the middleware.
-	middleware := jwtmiddleware.New(jwtValidator.ValidateToken)
-
-	http.ListenAndServe("0.0.0.0:3000", middleware.CheckJWT(handler))
-}
