+++ go.mod
@@ -8,7 +8,6 @@ require (
 	github.com/google/go-cmp v0.5.7
 	github.com/pkg/errors v0.9.1
 	github.com/stretchr/testify v1.7.1
-	gopkg.in/square/go-jose.v2 v2.6.0
 )
 
 require (
@@ -27,6 +26,7 @@ require (
 	github.com/ugorji/go/codec v1.1.7 // indirect
 	golang.org/x/crypto v0.0.0-20220315160706-3147a52a75dd // indirect
 	golang.org/x/sys v0.0.0-20210615035016-665e8c7367d1 // indirect
+	gopkg.in/square/go-jose.v2 v2.6.0 // indirect
 	gopkg.in/yaml.v2 v2.2.8 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
 )
