+++ option.go
@@ -1,5 +1,7 @@
 package jwtmiddleware
 
+import jwtmiddleware "github.com/auth0/go-jwt-middleware/v2"
+
 // Option is how options for the JWTMiddleware are set up.
 type Option func(*JWTMiddleware)
 
@@ -39,35 +41,8 @@ func WithErrorHandler(h ErrorHandler) Option {
 // the JWT to be validated from the request.
 //
 // Default value: AuthHeaderTokenExtractor.
-func WithTokenExtractor(e TokenExtractor) Option {
+func WithTokenExtractor(e jwtmiddleware.TokenExtractor) Option {
 	return func(m *JWTMiddleware) {
-		m.tokenExtractor = e
-	}
-}
-
-// Option is how options for the JWTMiddleware are set up.
-type GinOption func(*GinJWTMiddleware)
-
-func GinWithCredentialsOptional(value bool) GinOption {
-	return func(m *GinJWTMiddleware) {
-		m.credentialsOptional = value
-	}
-}
-
-func GinWithValidateOnOptions(value bool) GinOption {
-	return func(m *GinJWTMiddleware) {
-		m.validateOnOptions = value
-	}
-}
-
-func GinWithErrorHandler(h GinErrorHandler) GinOption {
-	return func(m *GinJWTMiddleware) {
-		m.errorHandler = h
-	}
-}
-
-func GinWithTokenExtractor(e TokenExtractor) GinOption {
-	return func(m *GinJWTMiddleware) {
 		m.tokenExtractor = e
 	}
 }
